pyproject.toml
[project]
name = "story-assist"
version = "0.1.0"
description = "Story sketch expansion assistant with multi-provider LLM search"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic==0.84.0",
    "fastapi==0.135.1",
    "google-genai==1.65.0",
    "openai==2.24.0",
    "python-dotenv==1.2.2",
    "uvicorn[standard]==0.41.0",
]

[dependency-groups]
dev = [
    "httpx==0.28.1",
    "pytest==9.0.2",
    "pytest-asyncio==1.3.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
